requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-crispy-forms==1.9.2
django-etc==1.2.0
django-hitcount==1.3.3
django-mathfilters==1.0.0
django-taggit==1.3.0
django-tinymce==3.1.0
idna==2.10
Pillow==8.0.1
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
